feat: Remove default Node.js support
BREAKING CHANGE: Default Node.js support is removed. It is hard to use
different APIs in different environments right in today's bundled world.

Closes #116.
