+++ src/entropyProvider.ts
@@ -1,5 +1,5 @@
 import { UnsignedTypedArray } from './unsignedTypedArray';
 
 export interface EntropyProvider {
-    getRandomValues<T extends UnsignedTypedArray>(array: T): Promise<T>;
+    getRandomValues<T extends UnsignedTypedArray>(array: T): T | Promise<T>;
 }
