--- src/environmentDetectingEntropyProvider.ts
@@ -1,125 +1,0 @@
-import { EntropyProvider } from './entropyProvider';
-import { UnsignedTypedArray } from './unsignedTypedArray';
-
-export class EnvironmentDetectingEntropyProvider implements EntropyProvider {
-    /**
-     * The crypto implementation used in the browser. The window.crypto object is used.
-     */
-    private readonly browserCrypto?: Crypto;
-
-    /**
-     * The crypto implementation used in Node.js environments. This will be provided by the "crypto" module of Node.
-     */
-    private readonly nodeCrypto?: {
-        randomFill: <T extends UnsignedTypedArray>(buffer: T, callback: (err: Error | null, buf: T) => void) => void;
-    };
-
-    /**
-     * According to the Web Crypto standard, there is a 2 ** 16 bytes quota for requesting entropy.
-     */
-    private static readonly BROWSER_ENTROPY_QUOTA_BYTES = 65536;
-
-    /**
-     * The environment in which the package is run.
-     */
-    private readonly environment: 'browser' | 'node';
-
-    public constructor() {
-        if (typeof window !== 'undefined' && typeof window.document !== 'undefined') {
-            this.environment = 'browser';
-            if (typeof window.crypto !== 'undefined' && typeof window.crypto.getRandomValues !== 'undefined') {
-                this.browserCrypto = window.crypto;
-            } else {
-                throw new Error('window.crypto.getRandomValues not available');
-            }
-        } else if (
-            typeof process !== 'undefined' &&
-            typeof process.versions !== undefined &&
-            typeof process.versions.node !== undefined
-        ) {
-            this.environment = 'node';
-            try {
-                this.nodeCrypto = require('crypto');
-            } catch (e) {
-                throw new Error('NodeJS.crypto not available');
-            }
-        } else {
-            throw new Error('Unexpected environment: neither browser nor node');
-        }
-    }
-
-    /**
-     * Puts random values into the given @param array, and returns the array.
-     */
-    public async getRandomValues<T extends UnsignedTypedArray>(array: T): Promise<T> {
-        switch (this.environment) {
-            case 'browser':
-                return this.getRandomValuesBrowser(array);
-            case 'node':
-                return this.getRandomValuesNode(array);
-        }
-
-        throw new Error('AssertError: unexpected environment in getRandomValues');
-    }
-
-    /**
-     * Puts random values into the given @param array in a browser environment, and returns the array.
-     * If the array's length is greater than the general @member BROWSER_ENTROPY_QUOTA_BYTES,
-     * it is divided into chunks, and filled chunk-by-chunk.
-     * Can be only called, if @member environment is 'browser'.
-     */
-    private getRandomValuesBrowser<T extends UnsignedTypedArray>(array: T): T {
-        if (this.environment !== 'browser') {
-            throw new Error('AssertError: not in browser environment');
-        }
-
-        if (this.browserCrypto === undefined) {
-            throw new Error('AssertError: no browserCrypto in browser environment');
-        }
-
-        if (array.byteLength <= EnvironmentDetectingEntropyProvider.BROWSER_ENTROPY_QUOTA_BYTES) {
-            return this.browserCrypto.getRandomValues(array);
-        }
-
-        let remainingBytes = array.byteLength;
-        while (remainingBytes > 0) {
-            const availableEntropyBytes = Math.min(
-                remainingBytes,
-                EnvironmentDetectingEntropyProvider.BROWSER_ENTROPY_QUOTA_BYTES,
-            );
-            const chunkStart = array.byteLength - remainingBytes;
-            const chunkLength = availableEntropyBytes / array.BYTES_PER_ELEMENT;
-            const chunkEnd = chunkStart + chunkLength;
-            const chunkToFill = array.subarray(chunkStart, chunkEnd);
-            this.browserCrypto.getRandomValues(chunkToFill);
-            remainingBytes -= availableEntropyBytes;
-        }
-
-        return array;
-    }
-
-    /**
-     * Puts random values into the given @param array in a Node.js environment, and returns the array.
-     * Can be only called, if @member environment is 'node'.
-     */
-    private async getRandomValuesNode<T extends UnsignedTypedArray>(array: T): Promise<T> {
-        if (this.environment !== 'node') {
-            throw new Error('not in node environment');
-        }
-
-        return new Promise<T>((resolve, reject): void => {
-            if (this.nodeCrypto === undefined) {
-                throw new Error('AssertError: no nodeCrypto in node environment');
-            }
-
-            this.nodeCrypto.randomFill(array, (error: Error | null, array: T) => {
-                if (error !== null) {
-                    reject(error);
-                    return;
-                }
-
-                resolve(array);
-            });
-        });
-    }
-}
