+++ test/utils/windowMock.ts
@@ -1,0 +1,7 @@
+import { randomFillSync } from 'crypto';
+
+export const windowMock = () => ({
+    crypto: {
+        getRandomValues: (array: Uint8Array) => randomFillSync(array),
+    },
+});
